feat: themeSwitch 컴포넌트 모드에 따른 이미지 변경
